+++ tests/fixture_plugins/pacbio_fixtures/name_fixtures.py
@@ -3,6 +3,8 @@
 import pytest
 
 from cg.constants.pacbio import PacBioDirsAndFiles
+from cg.services.post_processing.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
+from cg.services.post_processing.pacbio.metrics_parser.models import PacBioMetrics
 
 
 @pytest.fixture
@@ -26,3 +28,19 @@ def pac_bio_1_a01_cell_full_name() -> str:
 def ccs_report_1_a01_name(pac_bio_1_a01_cell_full_name: str) -> str:
     """Return the name of a ccs report file."""
     return f"{pac_bio_1_a01_cell_full_name}.{PacBioDirsAndFiles.CCS_REPORT_SUFFIX}"
+
+
+@pytest.fixture
+def expected_smrt_cell_bundle_name(
+    pac_bio_metrics_parser: PacBioMetricsParser, expected_pac_bio_run_data
+) -> str:
+    parsed_metrics: PacBioMetrics = pac_bio_metrics_parser.parse_metrics(expected_pac_bio_run_data)
+    return parsed_metrics.dataset_metrics.cell_id
+
+
+@pytest.fixture
+def expexted_pac_bio_sample_name(
+    pac_bio_metrics_parser: PacBioMetricsParser, expected_pac_bio_run_data
+):
+    parsed_metrics: PacBioMetrics = pac_bio_metrics_parser.parse_metrics(expected_pac_bio_run_data)
+    return parsed_metrics.dataset_metrics.sample_internal_id
