+++ tests/services/post_processing/pacbio/conftest.py
@@ -26,42 +26,3 @@ def expected_pac_bio_run_data(
         well_name="A01",
         plate="1",
     )
-
-
-@pytest.fixture
-def pac_bio_run_file_manager() -> PacBioRunFileManager:
-    return PacBioRunFileManager()
-
-
-@pytest.fixture
-def pac_bio_metrics_parser(pac_bio_run_file_manager: PacBioRunFileManager) -> PacBioMetricsParser:
-    return PacBioMetricsParser(file_manager=pac_bio_run_file_manager)
-
-
-@pytest.fixture
-def pac_bio_housekeeper_service(
-    real_housekeeper_api: HousekeeperAPI,
-    pac_bio_run_file_manager: PacBioRunFileManager,
-    pac_bio_metrics_parser: PacBioMetricsParser,
-) -> PacBioHousekeeperService:
-    return PacBioHousekeeperService(
-        hk_api=real_housekeeper_api,
-        file_manager=pac_bio_run_file_manager,
-        metrics_parser=pac_bio_metrics_parser,
-    )
-
-
-@pytest.fixture
-def expected_smrt_cell_bundle_name(
-    pac_bio_metrics_parser: PacBioMetricsParser, expected_pac_bio_run_data
-) -> str:
-    parsed_metrics: PacBioMetrics = pac_bio_metrics_parser.parse_metrics(expected_pac_bio_run_data)
-    return parsed_metrics.dataset_metrics.cell_id
-
-
-@pytest.fixture
-def expexted_pac_bio_sample_name(
-    pac_bio_metrics_parser: PacBioMetricsParser, expected_pac_bio_run_data
-):
-    parsed_metrics: PacBioMetrics = pac_bio_metrics_parser.parse_metrics(expected_pac_bio_run_data)
-    return parsed_metrics.dataset_metrics.sample_internal_id
